requirements.txt
Pillow
matplotlib
pytest
scikit-image

